requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
jwt==0.6.1
paramiko==2.6.0
PyJWT==1.7.1
PyMySQL==0.9.3